feat: sendMessage (as transfers)
right now this implements messaging as transfers that the receiver cannot fulfill. that is not a good long-term solution because it is slow and spams the ledger. it does, however, have the advantage of enabling any account on the ledger to send to any other account (which is not the case with something like HTTP RPC)